requirements.txt
opencv-python
numpy
psycopg2-binary
python-dotenv
pyserial
flask
flask-socketio
plotly 